+++ CHANGELOG.md
@@ -7,6 +7,7 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 ## Unreleased
 ### Changed
 - Use the new [Conch](https://github.com/isambard-sc/conch/) CA
+- Only write the SSH config if the contents has actually changed.
 
 ## [0.1.4] - 2024-09-10
 ### Added